Several changes in favor of organization
Renamed header files to reflect their use -> page_frame.php and %v2.php. Since the headers also include their own footer it is more appropriately named page_frame.\n Got rid of the Home option in the header in exchange for allowing the Zoo logo to be pressed to redirect back to home (See salisburyzoo.org)\n Changed Edit Welfare Forms to Welfare Forms in the header because I aim to add a button in the forms to allow a zookeeper to start a checkup from there and choose the animals they are examining.